+++ LibSvm/SolverNu.cs
@@ -64,8 +64,8 @@ namespace LibSvm
 
       int ip = Gmaxp_idx;
       int @in = Gmaxn_idx;
-      float[] Q_ip = null;
-      float[] Q_in = null;
+      double[] Q_ip = null;
+      double[] Q_in = null;
       if (ip != -1) // null Q_ip not accessed: Gmaxp=-INF if ip=-1
         Q_ip = _q.GetQ(ip, _activeSize);
       if (@in != -1)
